# perf: optimize scale finding algorithm from O(n) to O(1)