--- a/Cargo.toml
+++ b/Cargo.toml
@@ -25,7 +25,14 @@ rayon = "1.11.0"
 uuid = { version = "1.18.0", features = ["v4"] }
 thiserror = { version = "2.0.16" }
 
+[dev-dependencies]
+criterion = "0.5"
+
 [profile.release]
 lto = true
 opt-level = 3
 codegen-units = 1
+
+[[bench]]
+name = "scale_finding"
+harness = false