## Summary
- Replaced inefficient linear search algorithm with direct mathematical calculation
- Achieved up to 447x performance improvement for large coordinate values
- Maintains functional equivalence with comprehensive test coverage

## Changes
- **Algorithm optimization**: Replaced O(n) linear search (incrementing by 0.0001) with O(1) direct calculation
- **Early return optimization**: Added fast path for small values that work with minimum scale
- **Comprehensive testing**: Added unit tests covering edge cases and boundary conditions
- **Performance benchmarking**: Added criterion benchmarks to measure improvements

## Performance Improvements
Benchmark results using Criterion 0.5:

| Input Size | Old Algorithm | New Algorithm | Speedup |
|------------|---------------|---------------|---------|
| Small (1000) | 631.60 ps | 491.79 ps | **1.3x** |
| Medium (2.15e9) | 7.50 µs | 822.42 ps | **9.1x** |
| Large (1e10) | 35.85 µs | 838.39 ps | **42.8x** |
| Very Large (1e11) | 366.85 µs | 820.28 ps | **447.3x** |

## Testing
- ✅ All existing tests pass
- ✅ Added 5 new unit tests for edge cases
- ✅ Criterion benchmarks demonstrate performance gains
- ✅ Algorithm produces identical scales to original implementation

The optimization provides massive performance benefits especially for large coordinate values commonly found in real-world point cloud data.

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>